Implemented displaying Offer and Answer data as QR codes for easier copying
This change adds the ability to display Offer and Answer data as QR codes. When a user clicks on an Offer or Answer in the console, a dialog will appear showing the data as a QR code. This makes it easier to copy the data, especially when dealing with long strings.

The following changes were made:
- Added a dependency on the ZXing library for QR code generation.
- Modified the ConsoleAdapter to handle click events on Offer and Answer messages.
- Implemented a dialog to display the QR code.